When closing consumer also delete from hashmap
Because of the way kafkaconsumer objects are stored and accessed through
KafkaUtils we need to delete those objects upon connection close,
otherwise if we get the object using zkconnectionurl -string we get an
consumer object that has already been closed.

Using an object that is already closed will cause
java.lang.IllegalStateException and that's of course not a perfect
situation. This change also allows using the same zk connectstring for
managed cluster and kafkastats, which especially makes development
easier because you don't have to setup two clusters.

A better fix would probably be to inherit KafkaConsumer and make the
close to make sure that the object gets alos deleted from everywhere
else, but this is sufficient enough at this point.